+++ app/user/signin/page.tsx
@@ -7,46 +7,11 @@ import Button from '@/components/ui/Button';
 import Icon from '@/components/shared/Icon';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { signinUser } from '@/api/auth';
-import { useMutation } from '@tanstack/react-query';
-import { toast } from 'react-toastify';
-import Cookies from 'js-cookie';
-import { jwtDecode } from 'jwt-decode';
+
+import { useSignin } from '@/hooks/useAuth';
 
 export default function Signin() {
-  const router = useRouter();
-
-  const { mutate } = useMutation({
-    mutationFn: signinUser,
-    onMutate: async (data) => {
-      console.log('제출된 로그인 데이터:', data);
-    },
-    onSuccess: (response) => {
-      console.log(' 로그인 데이터:', response);
-      const accessToken = response.accessToken;
-      if (accessToken) {
-        //로컬 스토리지에 토큰 저장
-        localStorage.setItem('accessToken', accessToken);
-        console.log('accessToken 로컬스토리지 저장 완료:', localStorage.getItem('accessToken'));
-        // 쿠키에 토큰 저장 (만료시간 1일 설정)
-        Cookies.set('accessToken', accessToken, { expires: 1 });
-        console.log('accessToken 쿠키에 저장 완료:', Cookies.get('accessToken'));
-        // 토큰 디코딩 test
-        const decodedToken: any = jwtDecode(accessToken);
-        console.log('디코딩된 유저 정보:', decodedToken);
-
-        toast.success('성공적으로 로그인 되었습니다 :)', {
-          hideProgressBar: true,
-          autoClose: 900,
-        });
-        router.push('/');
-      }
-    },
-    onError: (error: any) => {
-      console.error('로그인 실패:', error.response?.data || error.message);
-      toast.error('로그인에 실패했습니다.');
-    },
-  });
+  const { mutate } = useSignin();
 
   const handleSignin = (data: SignInRequestData) => {
     mutate(data);
