+++ components/layout/Gnb.tsx
@@ -11,9 +11,7 @@ import Icon from '../shared/Icon';
 import DropDown from '../ui/DropDown';
 
 import { useRouter } from 'next/navigation';
-import { toast } from 'react-toastify';
-import { signoutUser } from '@/api/auth';
-import { useMutation } from '@tanstack/react-query';
+import { useSignout } from '@/hooks/useAuth';
 
 function NavItem({
   href,
@@ -42,6 +40,7 @@ export default function GNB() {
   const pathname = usePathname();
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const router = useRouter();
+  const { mutate: logout } = useSignout();
 
   //로그인 여부 체크하기
   useEffect(() => {
@@ -51,31 +50,13 @@ export default function GNB() {
     }
   }, []);
 
-  // 로그아웃
-  const { mutate: logout } = useMutation({
-    mutationFn: signoutUser,
-    onSuccess: () => {
-      setIsLoggedIn(false);
-      toast.success('로그아웃 되었습니다', {
-        hideProgressBar: true,
-        autoClose: 900,
-      });
-      router.push('/');
-    },
-    onError: (error) => {
-      toast.error('로그아웃 실패', {
-        hideProgressBar: true,
-        autoClose: 900,
-      });
-    },
-  });
-
   // 드롭다운 아이템 Click시 handler
   const handleDropDownItem = (item: string) => {
     if (item === '마이페이지') {
       router.push('/myprofile'); // 마이페이지로 이동
     } else if (item === '로그아웃') {
       logout(); // 로그아웃 처리
+      setIsLoggedIn(false);
     }
   };
